requirements.txt
anyio==3.6.1
click==8.1.3
dnspython==2.2.1
email-validator==1.3.0
fastapi==0.85.1
h11==0.14.0
idna==3.4
motor==3.0.0
pydantic==1.10.2
PyJWT==2.6.0
pymongo==4.2.0
python-decouple==3.6
sniffio==1.3.0
starlette==0.20.4
typing_extensions==4.4.0
uvicorn==0.18.3
